[Decomp][SpellMgr] Cohesion split: extract proc data loaders to SpellMgrProc.cpp
